Add support for client comment and timestamp fields.

+++ scripts/pi-hole/php/groups.php
@@ -253,6 +253,28 @@ if ($_POST['action'] == 'get_groups') {
 } elseif ($_POST['action'] == 'edit_client') {
     // Edit client identified by ID
     try {
+        $stmt = $db->prepare('UPDATE client SET comment=:comment WHERE id = :id');
+        if (!$stmt) {
+            throw new Exception('While preparing statement: ' . $db->lastErrorMsg());
+        }
+
+        $comment = $_POST['comment'];
+        if (strlen($comment) == 0) {
+                // Store NULL in database for empty comments
+                $comment = null;
+        }
+        if (!$stmt->bindValue(':comment', $comment, SQLITE3_TEXT)) {
+            throw new Exception('While binding comment: ' . $db->lastErrorMsg());
+        }
+
+        if (!$stmt->bindValue(':id', intval($_POST['id']), SQLITE3_INTEGER)) {
+            throw new Exception('While binding id: ' . $db->lastErrorMsg());
+        }
+
+        if (!$stmt->execute()) {
+            throw new Exception('While executing: ' . $db->lastErrorMsg());
+        }
+
         $stmt = $db->prepare('DELETE FROM client_by_group WHERE client_id = :id');
         if (!$stmt) {
             throw new Exception('While preparing DELETE statement: ' . $db->lastErrorMsg());
